WIP: Use turbo for .js-delete – but it can't be done ☹️
turbo.js empties out the search part of the url[0], and there is no
way to override that behaviour. I was hoping to do some trickery to
change the url in the `turbo:submit-start` event, but it seems to call
the old url no matter what I try (and in any case we probably
shouldn't rely on internals, *very* hacky).

So instead, it's probably best to keep our `submitForm` for now, for
`.js-delete` support, and instead have a slow deprecation process
where we recommend people switch to turbo's delete method (and mention
the caveat that if you use search/query params in a <a href> they will
be changed to body form params).

[0] https://github.com/hotwired/turbo/commit/f8b0e17a0c605080de0097112a5cf408abef9bb1